dbt_airflow/pyproject.toml
[tool.poetry]
name = "airflow-dbt"
version = "0.1.0"
description = ""
authors = ["Asafe Felipe <asafe.santos@acerto.com.br>"]
readme = "README.md"
packages = [{include = "airflow_dbt"}]

[tool.poetry.dependencies]
python = "^3.11"
apache-airflow = "^2.6.2"
apache-airflow-providers-postgres = "^5.5.0"
dbt-postgres = "^1.5.1"
pandas = "^2.0.2"
astronomer-cosmos = "^0.7.3"
yfinance = "^0.2.22"
pandas-datareader = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
